Provide the prefix, working_dir and arch to the implicit runner in wineexec.
There's now a way to override these settings, but we only use it when
wineexec creates a runner. This runner has no config, and would fall
back on the default (system!) prefix without this commit.

That's not nice at all can't be controlled by the user.

The main effect of this commit is that DLLs and registry keys get
applied to the game's prefix before running a WINE installer. I think
this is more likely to make the installer work, but in any case
reconfiguring the system default wineprefix (~/.wine) to install in
some other prefix is quite antisocial.